+++ src/docs/data/builders/select.ts
@@ -42,7 +42,7 @@ const builder = builderSchema(BUILDER_NAME, {
 			see: SEE.BRING_YOUR_OWN_STORE,
 		},
 		{
-			name: 'onSelectChange',
+			name: 'onSelectedChange',
 			type: 'ChangeFn<SelectOption<unknown>>',
 			description: 'A callback that is called when the selected option changes.',
 			see: SEE.CHANGE_FUNCTIONS,
@@ -79,8 +79,8 @@ const builder = builderSchema(BUILDER_NAME, {
 			description: 'The builder store used to create the select options.',
 		},
 		{
-			name: 'input',
-			description: 'The builder store used to create the select input.',
+			name: 'hidden-input',
+			description: 'The builder store used to create the select hidden input.',
 		},
 		{
 			name: 'label',
@@ -190,12 +190,12 @@ const option = elementSchema('option', {
 	events: selectEvents['option'],
 });
 
-const input = elementSchema('input', {
+const hiddenInput = elementSchema('hidden-input', {
 	description: 'The hidden input element. Used for form submission.',
 	dataAttributes: [
 		{
-			name: 'data-melt-select-input',
-			value: ATTRS.MELT('input'),
+			name: 'data-melt-select-hidden-input',
+			value: ATTRS.MELT('hidden-input'),
 		},
 	],
 });
@@ -296,7 +296,18 @@ const keyboard: KeyboardSchema = [
 	},
 ];
 
-const schemas = [builder, trigger, label, menu, option, group, groupLabel, input, separator, arrow];
+const schemas = [
+	builder,
+	trigger,
+	label,
+	menu,
+	option,
+	group,
+	groupLabel,
+	separator,
+	arrow,
+	hiddenInput,
+];
 
 const features = [
 	'Full keyboard navigation',