+++ src/docs/previews/combobox/multi/tailwind/index.svelte
@@ -3,15 +3,6 @@
 	import { Check, ChevronDown, ChevronUp } from 'lucide-svelte';
 	import { fly } from 'svelte/transition';
 
-	const {
-		elements: { menu, input, option, label },
-		states: { open, inputValue, touchedInput },
-		helpers: { isSelected },
-	} = createCombobox({
-		forceVisible: true,
-		multiple: true,
-	});
-
 	type Manga = {
 		author: string;
 		title: string;
@@ -70,6 +61,15 @@
 			disabled: false,
 		},
 	];
+
+	const {
+		elements: { menu, input, option, label },
+		states: { open, inputValue, touchedInput },
+		helpers: { isSelected },
+	} = createCombobox({
+		forceVisible: true,
+		multiple: true,
+	});
 
 	$: filteredMangas = $touchedInput
 		? mangas.filter(({ title, author }) => {