+++ src/docs/previews/select/main/tailwind/index.svelte
@@ -53,9 +53,10 @@
 					{#each arr as item}
 						<div
 							class="relative cursor-pointer rounded-lg py-1 pl-8 pr-4 text-neutral-800
-							focus:z-10 focus:text-magnum-700
-						data-[highlighted]:bg-magnum-50 data-[selected]:bg-magnum-100
-						data-[highlighted]:text-magnum-900 data-[selected]:text-magnum-900"
+							hover:bg-magnum-100 focus:z-10
+							focus:text-magnum-700
+							data-[highlighted]:bg-magnum-200 data-[highlighted]:text-magnum-900
+							data-[disabled]:opacity-50"
 							use:melt={$option({ value: item, label: item })}
 						>
 							<div class="check {$isSelected(item) ? 'block' : 'hidden'}">