+++ src/lib/internal/actions/melt/index.ts
@@ -19,10 +19,11 @@ type GetActionAttributes<Builder> = Builder extends Record<string, any> & {
  * @example
  * ```svelte
  * <script>
- * 	const { builder, melt } = createBuilder();
+ * 	import { createLabel, melt } from '@melt-ui/svelte';
+ * 	const { elements: { root } } = createLabel();
  * </script>
  *
- * <div use:melt={$builder} />
+ * <label use:melt={$root} />
  * ```
  */
 export function melt<
@@ -32,6 +33,7 @@ export function melt<
 	A extends Record<string, any>,
 	Param = never
 >(node: Element, params: Builder): ActionReturn<Builder, Attributes> {
-	// @ts-expect-error calls the action for debugging purposes
-	return params.action(node);
+	throw new Error(
+		"[MELTUI ERROR]: The `use:melt` action cannot be used without MeltUI's Preprocessor. See: https://www.melt-ui.com/docs/preprocessor"
+	);
 }